requirements.txt
apscheduler==3.11.0
flask==2.3.3
flask-caching==2.1.0
Flask-Limiter==3.6.0
Flask-Redis==0.4.0
flask-sqlalchemy==3.1.1
gunicorn==22.0.0
mysql-connector-python==8.4.0
pymysql==1.1.1
redis==5.0.8
requests==2.32.5
tenacity==8.5.0
